Better way of fixing #46, #47 and #48
Previous fix is to find all the use of method's Body field and add a
check for whether the body needs to be processed before processing
method body.  A better way is to change Body to a property and add the
check to the get of the property value so that we won't miss the use
spot that need the check, also later useage of method body doesn't need
to be concerned of adding the check.